+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # The `zksolc` changelog
 
+## [Unreleased]
+
+### Changed
+
+- The warning about the default codegen is made more verbose and informative
+
 ## [1.5.11] - 2025-01-27
 
 ### Fixed